requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
crispy-bootstrap5==2025.4
Django==5.2.1
django-allauth==65.8.1
django-bootstrap5==25.1
django-crispy-forms==2.4
django-taggit==6.1.0
django-tinymce==4.1.0
docopt==0.6.2
idna==3.10
packaging==25.0
pillow==11.2.1
pipreqs==0.4.13
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
yarg==0.1.10
